requirements.txt
Django==4.2.17
mysqlclient
djangorestframework
drf-yasg